--- excelMicroservice/src/main/java/com/shirayev/excel/processing/client/statistics/IStatisticsClient.java
@@ -1,9 +1,0 @@
-package com.shirayev.excel.processing.client.statistics;
-
-import com.shirayev.excel.processing.entities.File;
-
-public interface IStatisticsClient {
-
-    void handlerTransferData(File file);
-
-}
