+++ statisticsPeoplePassage/src/main/java/com/shirayev/statistics/people/passage/servicies/implementation/StatisticsServiceImp.java
@@ -1,8 +1,8 @@
 package com.shirayev.statistics.people.passage.servicies.implementation;
 
 
-import com.shirayev.statistics.people.passage.servicies.IStatisticsPeoplePassageService;
-import com.shirayev.statistics.people.passage.servicies.IStatisticsService;
+import com.shirayev.statistics.people.passage.servicies.StatisticsPeoplePassageService;
+import com.shirayev.statistics.people.passage.servicies.StatisticsService;
 import com.shirayev.statistics.people.passage.model.CountPeoplePassageByAction;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,9 +12,9 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class StatisticsService implements IStatisticsService {
+public class StatisticsServiceImp implements StatisticsService {
 
-    private final IStatisticsPeoplePassageService statisticsPeoplePassageService;
+    private final StatisticsPeoplePassageService statisticsPeoplePassageService;
 
     @Override
     public List<CountPeoplePassageByAction> handlerGetStatisticsByActionAndAge(Time start, Time end) {
